requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# WebSockets
websockets==11.0.3

# Database
motor==3.3.1  # Async MongoDB driver
pymongo==4.5.0

# Geospatial libraries
h3==3.7.6  # Uber's H3 hexagonal hierarchical geospatial indexing system
geojson==3.0.1
shapely==2.0.1
pyproj==3.6.0  # For coordinate transformations

# Data generation and manipulation
faker==19.4.0  # For generating fake data
numpy==1.26.0
pandas==2.1.1

# ML (for demand prediction)
scikit-learn==1.3.1
joblib==1.3.2

# Testing
pytest==7.4.2
httpx==0.25.0  # For async HTTP client in testing